nasty bugfixing

+++ src/main/groovy/com/cgz/routing/RoutingService.groovy
@@ -7,6 +7,6 @@ import com.cgz.geomath.Point
  */
 interface RoutingService {
 
-    long travelTimeInMinutes(Point origin, Point destination, TravelMode travelMode)
+    double travelTimeInMinutes(Point origin, Point destination, TravelMode travelMode)
 
 }
